+++ src/factories/gameboard.js
@@ -1,4 +1,5 @@
 import Ship from './ship';
+import { randomShip } from '../auxFunctions';
 
 const GameBoard = (testing) => {
   const ship = {
@@ -16,29 +17,6 @@ const GameBoard = (testing) => {
     hit: [],
   };
 
-  const reset = () => {
-    positions = {};
-    positionCoordinates = [];
-    attacks = {
-      mis: [],
-      hit: [],
-    };
-    createPositions();
-
-    const types = Object.keys(ship);
-    for (let i = 0; types.length > i; i += 1) {
-      ship[types[i]].reset();
-    }
-  };
-
-  if (testing) {
-    setPosition('battleship', createShipCoords('battleship', 0, 0, 'v'));
-    setPosition('submarine', createShipCoords('submarine', 1, 0, 'h'));
-    setPosition('carrier'), createShipCoords('carrier', 0, 0, 'h');
-  } else {
-    createPositions();
-  }
-
   const receiveAttack = (x, y) => {
     if (positions[x] == null || positions[x][y] == null) {
       attacks.mis.push([x, y]);
@@ -50,19 +28,59 @@ const GameBoard = (testing) => {
     return true;
   };
 
-  const missedAttacks = () => {
-    return attacks.mis;
-  };
+  const missedAttacks = () => attacks.mis;
 
   const allSunk = () => {
     const types = Object.keys(ship);
     for (let i = 0; types.length > i; i += 1) {
-      if (ship[types[i]].isSunk() == false) {
+      if (ship[types[i]].isSunk() === false) {
         return false;
       }
     }
     return true;
   };
+
+  function createShipCoords(shipType, x, y, direction) {
+    const { length } = ship[shipType];
+    const coords = [];
+    if (direction === 'v') {
+      for (let i = 0; length > i; i += 1) {
+        if (x >= 10 || y + i >= 10) {
+          return null;
+        }
+        if (positions[x] != null && positions[x][y + i] != null) {
+          return null;
+        }
+        coords.push({ x, y: y + i });
+      }
+      return coords;
+    }
+    if (direction === 'h') {
+      for (let i = 0; length > i; i += 1) {
+        if (x + i >= 10 || y >= 10) {
+          return null;
+        }
+        if (positions[x + i] != null && positions[x + i][y] != null) {
+          return null;
+        }
+        coords.push({ x: x + i, y });
+      }
+      return coords;
+    }
+    return null;
+  }
+
+  function setPosition(shipToSet, coords) {
+    if (coords == null) return null;
+    for (let i = 0; coords.length > i; i += 1) {
+      if (positions[coords[i].x] == null) {
+        positions[coords[i].x] = {};
+      }
+      const location = { shipType: shipToSet, index: i };
+      positions[coords[i].x][coords[i].y] = location;
+    }
+    return 1;
+  }
 
   function createPositions() {
     const types = Object.keys(ship);
@@ -81,46 +99,28 @@ const GameBoard = (testing) => {
     }
   }
 
-  function setPosition(shipToSet, coords) {
-    if (coords == null) return null;
-    for (let i = 0; coords.length > i; i += 1) {
-      if (positions[coords[i].x] == null) {
-        positions[coords[i].x] = {};
-      }
-      const location = { shipType: shipToSet, index: i };
-      positions[coords[i].x][coords[i].y] = location;
-    }
-    return 1;
+  if (testing) {
+    setPosition('battleship', createShipCoords('battleship', 0, 0, 'v'));
+    setPosition('submarine', createShipCoords('submarine', 1, 0, 'h'));
+    setPosition('carrier', createShipCoords('carrier', 0, 0, 'h'));
+  } else {
+    createPositions();
   }
 
-  function createShipCoords(shipType, x, y, direction) {
-    const { length } = ship[shipType];
-    const coords = [];
-    if (direction === 'v') {
-      for (let i = 0; length > i; i += 1) {
-        if (x >= 10 || y + i >= 10) {
-          return null;
-        }
-        if (positions[x] != null && positions[x][y + i] != null) {
-          return null;
-        }
-        coords.push({ x: x, y: y + i });
-      }
-      return coords;
+  const reset = () => {
+    positions = {};
+    positionCoordinates = [];
+    attacks = {
+      mis: [],
+      hit: [],
+    };
+    createPositions();
+
+    const types = Object.keys(ship);
+    for (let i = 0; types.length > i; i += 1) {
+      ship[types[i]].reset();
     }
-    if (direction == 'h') {
-      for (let i = 0; length > i; i += 1) {
-        if (x + i >= 10 || y >= 10) {
-          return null;
-        }
-        if (positions[x + i] != null && positions[x + i][y] != null) {
-          return null;
-        }
-        coords.push({ x: x + i, y: y });
-      }
-      return coords;
-    }
-  }
+  };
 
   return {
     positions,
@@ -131,16 +131,6 @@ const GameBoard = (testing) => {
     allSunk,
     reset,
   };
-};
-
-const randomShip = () => {
-  const direction = ['v', 'h'];
-  const X = Math.floor(Math.random() * (10 - 0 + 1)) + 0;
-  const Y = Math.floor(Math.random() * (10 - 0 + 1)) + 0;
-
-  const directionStr = direction[Math.floor(Math.random() * (1 - 0 + 1)) + 0];
-
-  return { X, Y, directionStr };
 };
 
 export default GameBoard;
